+++ Client/Client.Commands.cs
@@ -294,7 +294,7 @@ public partial class Client
             HelionLog.Info($"No map markers to remove.");
     }
 
-    [ConsoleCommand("findkeys", "Finds the next key in the map.")]
+    [ConsoleCommand("findkeys", "Finds all keys in the automap.")]
     private void CommandFindKey(ConsoleCommandEventArgs args)
     {
         if (m_layerManager.WorldLayer == null)
@@ -303,16 +303,25 @@ public partial class Client
         m_layerManager.WorldLayer.World.FindKeys();
     }
 
-    [ConsoleCommand("findkeylines", "Finds the next locked key line in the map.")]
+    [ConsoleCommand("findkeylines", "Marks all key lines the automap.")]
     private void CommandFindKeyLine(ConsoleCommandEventArgs args)
     {
         if (m_layerManager.WorldLayer == null)
             return;
 
-        m_layerManager.WorldLayer.World.FindKeyLines();
+        m_layerManager.WorldLayer.World.FindKeyLines(FindKeyLineOptions.None);
     }
 
-    [ConsoleCommand("findexits", "Finds the next exit line/sector in the map.")]
+    [ConsoleCommand("findinactivekeylines", "Marks all the key lines in the automap that haven't been activated.")]
+    private void CommandFindInactiveKeyLine(ConsoleCommandEventArgs args)
+    {
+        if (m_layerManager.WorldLayer == null)
+            return;
+
+        m_layerManager.WorldLayer.World.FindKeyLines(FindKeyLineOptions.Inactive);
+    }
+
+    [ConsoleCommand("findexits", "Marks all exit lines/sectors in the automap.")]
     private void CommandFindExit(ConsoleCommandEventArgs args)
     {
         if (m_layerManager.WorldLayer == null)
@@ -1085,7 +1094,7 @@ public partial class Client
                     break;
 
                 case LevelChangeType.EndGame:
-                    Intermission(world, () => GetLastEndGame(world));
+                    Intermission(world, () => GetEndGame(world));
                     break;
             }
         }
@@ -1095,23 +1104,14 @@ public partial class Client
         }
     }
 
-    private static FindMapResult GetLastEndGame(IWorld world)
+    private static FindMapResult GetEndGame(IWorld world)
     {
         var mapInfo = world.ArchiveCollection.MapInfo.MapInfo;
-        var maps = mapInfo.Maps.OrderByDescending(x => x.Cluster);
-        foreach (var map in maps)
+        var map = mapInfo.GetEpisodeEndGame(world.MapInfo);
+        if (map != null)
         {
-            if (map.EndGame != null)
-            {
-                SetEndGame(world, map.Next, map);
-                return FindMapResult.CreateEndGame(map.MapName);
-            }
-
-            if (MapInfo.EndGameMaps.Contains(map.Next))
-            {
-                SetEndGame(world, map.Next, map);
-                return FindMapResult.CreateEndGame(map.Next);
-            }
+            SetEndGame(world, map.Next, map);
+            return FindMapResult.CreateEndGame(map.MapName);
         }
 
         return FindMapResult.CreateError("No EndGame");
@@ -1284,10 +1284,10 @@ public partial class Client
         QueueLoadMap(mapInfoDef, null, null, e);
     }
 
-    private FindMapResult GetNextLevel(MapInfoDef mapDef) => 
+    private FindMapResult GetNextLevel(MapInfoDef mapDef) =>
         m_archiveCollection.Definitions.MapInfoDefinition.MapInfo.GetNextMap(mapDef);
 
-    private FindMapResult GetNextSecretLevel(MapInfoDef mapDef) => 
+    private FindMapResult GetNextSecretLevel(MapInfoDef mapDef) =>
         m_archiveCollection.Definitions.MapInfoDefinition.MapInfo.GetNextSecretMap(mapDef);
 
     private void ShowConsole()