+++ Core/World/FindKeyLineOptions.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Helion.World;
+
+[Flags]
+public enum FindKeyLineOptions
+{
+    None,
+    Inactive = 1
+}