+++ Core/World/WorldBase.cs
@@ -2556,11 +2556,11 @@ public abstract partial class WorldBase : IWorld
             HighlightAreas.Add(new HighlightArea(((Entity)m_findObjects[i]).Position, HighlightSize));
     }
 
-    public void FindKeyLines()
+    public void FindKeyLines(FindKeyLineOptions options)
     {
         m_findObjects.Clear();
         HighlightAreas.Clear();
-        MarkSpecials.FindKeyLines(this, m_findObjects);
+        MarkSpecials.FindKeyLines(this, m_findObjects, options);
         if (m_findObjects.Count == 0)
         {
             DisplayMessage("No keys found");
@@ -3078,6 +3078,10 @@ public abstract partial class WorldBase : IWorld
             case CheatType.ExitSecret:
                 ClearConsole?.Invoke(this, EventArgs.Empty);
                 ExitLevel(cheat.CheatType == CheatType.ExitSecret ? ExitLevelArgs.NextSecretMap() : ExitLevelArgs.NextMap());
+                break;
+            case CheatType.EndGame:
+                ClearConsole?.Invoke(this, EventArgs.Empty);
+                ExitLevel(ExitLevelArgs.EndGame());
                 break;
         }
     }